reflector: spilt code into two parts
